# metal: enable ggml-alloc