Make ggml-alloc compatible with concurrent dispatch.

Test on M1 Max with `n_ctx=512, n_natch=512`.
Memory usage of compute buffer:
| Model | Master | PR |
| --- | --- | --- |
| 33B Q4_0 | 472MB | 117MB |

Token generation:
| Model | Master | PR |
| --- | --- | --- |
| 33B Q4_0 | 65.4ms/tok | 62.5ms/tok |

The reduced memory comes from #2411, and the inference speedup comes from #2358.