@ggerganov 2023-08-16
The `perplexity` tool fails for 70B model both with and without metal:

```
make -j && ./perplexity -m models/70B-v2/ggml-model-q4_0.bin -f build/wikitext-2-raw/wiki.test.raw -gqa 8
```