@slaren 2023-08-16
This error is very surprising to me because the graph that the allocator uses to measure the buffer size should be identical to the graph used by the perplexity tool (`n_past=0`, `n_tokens=512`), so the allocations should be identical in both cases.