@slaren 2023-08-16
Ok, there is one possible difference. `llama_build_graph` has some calls to `ggml_allocr_alloc` that should use the metal `parse_seq`. However, by the time the measure graph is built, `parse_seq` is still not set. So the measure graph may need to be rebuilt after calling `ggml_allocr_set_parse_seq`.