@slaren 2023-08-16
There is a possibility that this could happen if the addresses of the weights are within this range:
https://github.com/ggerganov/llama.cpp/blob/13a746c6b963ab13818f4d401d5de9966f5af1f5/ggml-alloc.c#L278-L279