@lshzh-ww 2023-08-17
@ggerganov @slaren fixed in #2639 

Strange that it didn't trigger before. Anyway, now our allocator is more robust.